--- a/src/RigidBodyDynamics.jl
+++ b/src/RigidBodyDynamics.jl
@@ -5,7 +5,7 @@ module RigidBodyDynamics
 include("tree.jl")
 
 import Base: convert, zero, one, *, +, /, -, call, inv, get, findfirst, Random.rand, Random.rand!
-import Base: hcat, show, showcompact, isapprox, dot, cross, unsafe_copy!, Array, eltype, copy, exp, log
+import Base: hcat, show, showcompact, isapprox, dot, cross, unsafe_copy!, Array, eltype, copy, exp, log, get!
 using StaticArrays
 using Quaternions
 using DataStructures
@@ -24,7 +24,6 @@ include("cache_element.jl")
 
 importall .TreeDataStructure
 include("mechanism.jl")
-include("transform_cache.jl")
 include("mechanism_state.jl")
 include("mechanism_algorithms.jl")
 include("parse_urdf.jl")
@@ -112,7 +111,6 @@ export
     add_frame!,
     twist_wrt_world,
     relative_twist,
-    transform_to_parent,
     transform_to_root,
     relative_transform,
     mass,