--- a/src/cache_element.jl
+++ b/src/cache_element.jl
@@ -1,23 +1,31 @@
-type CacheElement{T, F}
-    data::T
-    updateFunction::F
+type CacheElement{T}
     dirty::Bool
-    constant::Bool
-    CacheElement(value::T, updateFunction::F, constant::Bool) = new(value, updateFunction, !constant, constant)
+    data::T
+    CacheElement() = new(true)
+    CacheElement(data::T) = new(true, data)
 end
-CacheElement{T, F}(::Type{T}, updateFunction::F) = CacheElement{T, F}(updateFunction(), updateFunction, false)
-CacheElement{T, F}(value::T, updateFunction::F) = CacheElement{T, F}(value, updateFunction, true)
-CacheElement{T}(value::T) = CacheElement{T, Function}(value, () -> value, true) # TODO: remove
 
-function get{T, F}(element::CacheElement{T, F})
-    if element.dirty
-        element.data = element.updateFunction()
-        element.dirty = false
-    end
+CacheElement{T}(data::T) = CacheElement{T}(data)
+
+function update!{T}(element::CacheElement{T}, data::T)
+    element.data = data
+    element.dirty = false
+end
+
+function get(element::CacheElement)
+    element.dirty && error("Cache dirty.")
     element.data
 end
+
 function setdirty!(element::CacheElement)
-    if !element.constant
-        element.dirty = true
+    element.dirty = true
+end
+
+@inline isdirty(element::CacheElement) = element.dirty
+
+function get!(element::CacheElement, updateFunction)
+    if isdirty(element)
+        update!(element, updateFunction())
     end
+    element.data
 end