Complete rewrite of `MechanismState` (and `CacheElement`); makes things a lot faster and results in cleaner code.

`MechanismState` now contains a `Tree{RigidBodyState, JointState}` that mirrors the `Tree{RigidBody, Joint}` in `Mechanism`. This results in much fewer map lookups in the recursive algorithms (mass matrix, inverse dynamics).

Also cleaned up/deduplicated a lot of code in `frames.jl` and `spatial.jl`.

Before:
```
  "inverse_dynamics" => BenchmarkTools.Trial: 
  memory estimate:  33.64 kb
  allocs estimate:  433
  --------------
  minimum time:     69.504 μs (0.00% GC)
  median time:      74.129 μs (0.00% GC)
  mean time:        78.622 μs (4.20% GC)
  maximum time:     2.840 ms (95.64% GC)

  "dynamics" => BenchmarkTools.Trial: 
  memory estimate:  45.05 kb
  allocs estimate:  617
  --------------
  minimum time:     121.035 μs (0.00% GC)
  median time:      123.758 μs (0.00% GC)
  mean time:        154.055 μs (3.18% GC)
  maximum time:     3.259 ms (91.73% GC)

  "mass_matrix" => BenchmarkTools.Trial: 
  memory estimate:  40.98 kb
  allocs estimate:  488
  --------------
  minimum time:     65.411 μs (0.00% GC)
  median time:      67.880 μs (0.00% GC)
  mean time:        72.183 μs (4.24% GC)
  maximum time:     1.986 ms (95.24% GC)
```

After:
```
  "inverse_dynamics" => BenchmarkTools.Trial: 
  memory estimate:  26.63 kb
  allocs estimate:  770
  --------------
  minimum time:     48.305 μs (0.00% GC)
  median time:      49.863 μs (0.00% GC)
  mean time:        52.395 μs (3.70% GC)
  maximum time:     1.836 ms (96.13% GC)

  "dynamics" => BenchmarkTools.Trial: 
  memory estimate:  31.13 kb
  allocs estimate:  1019
  --------------
  minimum time:     97.097 μs (0.00% GC)
  median time:      99.218 μs (0.00% GC)
  mean time:        118.370 μs (2.58% GC)
  maximum time:     2.737 ms (92.13% GC)

  "mass_matrix" => BenchmarkTools.Trial: 
  memory estimate:  21.69 kb
  allocs estimate:  578
  --------------
  minimum time:     43.209 μs (0.00% GC)
  median time:      44.550 μs (0.00% GC)
  mean time:        46.197 μs (2.19% GC)
  maximum time:     1.305 ms (95.39% GC)
```